rename motion_planning_frame_joints.* -> motion_planning_frame_joints_widget.*
