+++ moveit_ros/visualization/motion_planning_rviz_plugin/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(HEADERS
   include/moveit/motion_planning_rviz_plugin/motion_planning_display.h
   include/moveit/motion_planning_rviz_plugin/motion_planning_frame.h
-  include/moveit/motion_planning_rviz_plugin/motion_planning_frame_joints.h
+  include/moveit/motion_planning_rviz_plugin/motion_planning_frame_joints_widget.h
   include/moveit/motion_planning_rviz_plugin/motion_planning_param_widget.h
 )
 qt_wrap_ui(UIC_FILES
@@ -19,7 +19,7 @@ set(SOURCE_FILES
   src/motion_planning_frame_objects.cpp
   src/motion_planning_frame_scenes.cpp
   src/motion_planning_frame_states.cpp
-  src/motion_planning_frame_joints.cpp
+  src/motion_planning_frame_joints_widget.cpp
   src/motion_planning_display.cpp
   src/motion_planning_frame_manipulation.cpp
   src/motion_planning_param_widget.cpp
